# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Chatbot API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface OpenrouterConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface OpenrouterMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  /** @nullable */
  model: string | null;
  createdAt: string;
}

export interface OpenrouterConversationInput {
  title: string;
}

export interface OpenrouterMessageInput {
  content: string;
  model: string;
  agentMode?: boolean;
}

export interface OpenrouterConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: OpenrouterMessage[];
}

export interface OpenrouterError {
  error: string;
}

export interface AiModel {
  id: string;
  name: string;
  description: string;
  isFree: boolean;
}

export interface WebSearchInput {
  query: string;
  maxResults?: number;
}

export interface SearchResultItem {
  title: string;
  url: string;
  snippet: string;
}

export interface WebSearchResult {
  query: string;
  results: SearchResultItem[];
}

export interface FetchPageInput {
  url: string;
}

export interface FetchPageResult {
  url: string;
  title: string;
  content: string;
}