# lib/api-zod/src/generated/types/fetchPageResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Chatbot API specification
 * OpenAPI spec version: 0.1.0
 */

export interface FetchPageResult {
  url: string;
  title: string;
  content: string;
}