# lib/api-zod/src/generated/types/openrouterMessage.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Chatbot API specification
 * OpenAPI spec version: 0.1.0
 */

export interface OpenrouterMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  /** @nullable */
  model: string | null;
  createdAt: Date;
}