# lib/api-zod/src/generated/types/webSearchInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Chatbot API specification
 * OpenAPI spec version: 0.1.0
 */

export interface WebSearchInput {
  query: string;
  maxResults?: number;
}