# lib/api-zod/src/generated/types/webSearchResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Chatbot API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SearchResultItem } from './searchResultItem';

export interface WebSearchResult {
  query: string;
  results: SearchResultItem[];
}